ns: Refactor device registry validation
