+++ config/messages.json
@@ -8258,7 +8258,7 @@
     },
     "description": {
       "package": "pkg/networkserver",
-      "file": "grpc_deviceregistry.go"
+      "file": "device_state.go"
     }
   },
   "error:pkg/networkserver:field_not_zero": {
@@ -8267,7 +8267,7 @@
     },
     "description": {
       "package": "pkg/networkserver",
-      "file": "grpc_deviceregistry.go"
+      "file": "device_state.go"
     }
   },
   "error:pkg/networkserver:field_value": {
